+++ docs/examples/throttling/algorithm_customization.py
@@ -18,5 +18,4 @@ class SyncController(Controller[PydanticSerializer]):
 
 
 # run: {"controller": "SyncController", "method": "get", "url": "/api/sync/"}  # noqa: ERA001
-# run: {"controller": "SyncController", "method": "get", "url": "/api/sync/", "curl_args": ["-D", "-"], "assert-error-text": "Too many requests", "fail-with-body": false}  # noqa: ERA001, E501
 # openapi: {"controller": "SyncController", "openapi_url": "/docs/openapi.json"}  # noqa: ERA001