+++ docs/tools/sphinx_ext/run_examples.py
@@ -250,7 +250,7 @@ class _BaseBuilder:  # noqa: WPS214
         settings.configure(
             ROOT_URLCONF='url_conf',
             ALLOWED_HOSTS=['*'],
-            DEBUG=True,  # NOTE: this must be `False`
+            DEBUG=False,  # NOTE: this must be `False`
             SECRET_KEY='dummy-key-for-examples',  # noqa: S106
             INSTALLED_APPS=[
                 'django.contrib.auth',
@@ -1012,6 +1012,8 @@ class LiteralInclude(_LiteralInclude):  # noqa: WPS214
             )
         if openapi_result:
             nodes.append(_create_openapi_admonition(openapi_result))
+        if settings.DEBUG:
+            raise ValueError('settings.DEBUG is set, turn it off')
         return nodes
 
     def _generate_nodes(